Move family layout tests to the shared Rust API
