+++ crates/noon/tests/family_leaf_order_contract.rs
@@ -1,5 +1,5 @@
+use noon::{FamilyTranslation, Mobject};
 use noon_core::SemanticStore;
-use noon_web::{FrontendFamilyTranslation, Mobject};
 
 fn nested_family() -> (
     SemanticStore,
@@ -26,7 +26,7 @@ fn nested_family_translation_uses_authoritative_semantic_leaf_order() {
     let authoring_store =
         std::rc::Rc::new(std::cell::RefCell::new(noon_core::SemanticStore::new()));
     let (store, root, [first, second, third]) = nested_family();
-    let mut translation = FrontendFamilyTranslation::begin(&store, root, 0.25, -0.5).unwrap();
+    let mut translation = FamilyTranslation::begin(&store, root, 0.25, -0.5).unwrap();
     let mut first_handle =
         Mobject::manim_square(std::rc::Rc::clone(&authoring_store), 1.0).unwrap();
     let mut second_handle =
@@ -35,21 +35,21 @@ fn nested_family_translation_uses_authoritative_semantic_leaf_order() {
         Mobject::manim_square(std::rc::Rc::clone(&authoring_store), 1.0).unwrap();
 
     translation.apply(first, &mut first_handle).unwrap();
-    assert_eq!(first_handle.wire_translation().unwrap(), (0.25, -0.5));
+    assert_eq!(first_handle.center().unwrap(), (0.25, -0.5));
 
-    let second_before = second_handle.wire_translation().unwrap();
+    let second_before = second_handle.center().unwrap();
     let error = translation
         .apply(third, &mut second_handle)
         .expect_err("frontend wrapper order must not override semantic family order");
     assert!(error.contains("leaf mismatch"));
-    assert_eq!(second_handle.wire_translation().unwrap(), second_before);
+    assert_eq!(second_handle.center().unwrap(), second_before);
 
     translation.apply(second, &mut second_handle).unwrap();
     translation.apply(third, &mut third_handle).unwrap();
     translation.finish().unwrap();
 
-    assert_eq!(second_handle.wire_translation().unwrap(), (0.25, -0.5));
-    assert_eq!(third_handle.wire_translation().unwrap(), (0.25, -0.5));
+    assert_eq!(second_handle.center().unwrap(), (0.25, -0.5));
+    assert_eq!(third_handle.center().unwrap(), (0.25, -0.5));
 }
 
 #[test]
@@ -57,7 +57,7 @@ fn family_translation_refuses_incomplete_leaf_traversal() {
     let authoring_store =
         std::rc::Rc::new(std::cell::RefCell::new(noon_core::SemanticStore::new()));
     let (store, root, [first, _, _]) = nested_family();
-    let mut translation = FrontendFamilyTranslation::begin(&store, root, 1.0, 0.0).unwrap();
+    let mut translation = FamilyTranslation::begin(&store, root, 1.0, 0.0).unwrap();
     let mut handle = Mobject::manim_square(std::rc::Rc::clone(&authoring_store), 1.0).unwrap();
 
     translation.apply(first, &mut handle).unwrap();
